Add :Async option to detach connections and handle in separate actors
